batch_processing_service/requirements.txt
# Web framework
Flask==2.0.1
gunicorn==20.1.0

# Background jobs and queue
rq==1.10.1
redis==4.3.4

# Distributed computing
ray==2.2.0

# Workflow orchestration
prefect==2.7.7

# Environment and utilities
python-dotenv==0.19.1
requests==2.28.1

# Testing
pytest==7.0.1
pytest-flask==1.2.0
